Refactor button components
By using the built in React.ButtonHTMLAttributes<HTMLButtonElement>
type for these custom buttons props, there is no need to reimplement
those properties as every props of a default button will be accessible.

Default values for some props are defined as well which is optional to be
overrided by the user of the component.

There are also props that are defined with values unique to that component only,
so that the user of the component cannot alter those properties (e.g style, className).